Full implementation of logging system
This finishes the implementation of the logging system, enabling the users to
redirect the error messages to a file and to only see the amount of information
specified. By default, Warnings and Errors are shown.

If the user is on SimpleCV's interactive shell, the uncaught exceptions (errors)
are displayed just like on any IPython shell, unless logging to file is enabled.