Fixing the Image operators(__add__,__sub__,__and__,__or__) for scalars.
They weren't working as intended with color pictures and scalars. Example:
Image((100,100))+100 should yield an image that's gray now. Before it only added to the blue channel.

I guess there's probably a need for adding tuples to the image, but I didn't really need that atm. Though i still think this is a better behavoir for the
moment(adding to all channels).