Create motion_detection_SimplCV.py
A Simple program in python for motion detection and drawing lines over the moving Object. Here we accepts two images of different time and substract them . The Sucbstracted Image holds the key . The image show the object in motion everything else is substracted off.
Lines are found on substracted Image and drawn to the normal image.