Changes requirements from PIL to pillow
As requested on #468, pillow has several advantages over PIL and this change
should enable getting simplecv to Fedora's package system.

I uninstalled PIL, installed pillow and had no problems whatsoever, but I can't
assure that'll be the case for everyone. Although nobody needs to actually do
that for now, it'd be nice if people with other OSs try that and run the tests,
making sure this transition is indeed as smooth as promised.

Notice that this commit has no changes on code whatsoever: We already import PIL
the way pillow is imported.